PnL and details card fix
