+++ features/earn/guni/manage/containers/GuniManageMultiplyVaultDetails.tsx
@@ -1,7 +1,6 @@
 import {
   AfterPillProps,
   getAfterPillColors,
-  VaultDetailsCardNetValue,
   VaultDetailsSummaryContainer,
   VaultDetailsSummaryItem,
 } from 'components/vault/VaultDetails'
@@ -11,6 +10,7 @@ import { useTranslation } from 'next-i18next'
 import React from 'react'
 import { Box, Grid } from 'theme-ui'
 
+import { VaultDetailsCardNetValue } from '../../../../../components/vault/detailsCards/VaultDetailsCardNetValue'
 import { ManageMultiplyVaultState } from '../../../../multiply/manage/pipes/manageMultiplyVault'
 
 function GuniManageMultiplyVaultDetailsSummary({
